Add vud-field sub-module.
This field relies on core field system to create a field which adds a
Vote Up/Down voting widget to the any entity.

This solves the per content type, per content configuration, storing
widget value on the field value. Maybe we can add more information there
in the future, but this is the start.